refactor: Car 생성을 다른 객체에 위임
