+++ src/main/java/racing/domain/RacingHost.java
@@ -1,0 +1,9 @@
+package racing.domain;
+
+import java.util.List;
+
+public interface RacingHost {
+
+    Racing holdRacing(List<String> participants, int round);
+
+}
