+++ src/test/java/racing/domain/RacingTest.java
@@ -1,11 +1,7 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.domain.Car;
-import racing.domain.Racing;
-import racing.domain.ReadOnlyCarProxy;
-import racing.domain.ReadonlyCar;
 import racing.dto.RacingReport;
 
 import java.util.Arrays;
